Change delete element logic
Now on click on trash we change row element style for set it unclickable
and get it look like we did something with element (with danger class
style)

We send delete request on server and server not tell us to redirect but
just answer by code 200 or 404. To tell us what to do with element
If we receive 404 - we return styles to previously state
If we receive 200 - we set element display style to none.

iss: #2174